+++ ChangeLog.md
@@ -1,3 +1,7 @@
+2018-12-19 0.0.8
+
+- rename project to 'pxs' (from 'pxs-ui')
+
 2018-12-14 0.0.8
 
 - initial preparations to edit geometries